+++ spec/policies/activity_policy_spec.rb
@@ -19,7 +19,7 @@ RSpec.describe ActivityPolicy do
         is_expected.to permit_action(:create)
         is_expected.to permit_action(:edit)
         is_expected.to permit_action(:update)
-        is_expected.to permit_action(:download)
+        is_expected.to forbid_action(:show_xml)
         is_expected.to forbid_action(:destroy)
         is_expected.to forbid_action(:redact_from_iati)
         is_expected.to forbid_action(:create_refund)
@@ -41,8 +41,8 @@ RSpec.describe ActivityPolicy do
         is_expected.to permit_action(:create)
         is_expected.to permit_action(:edit)
         is_expected.to permit_action(:update)
-        is_expected.to permit_action(:download)
 
+        is_expected.to forbid_action(:show_xml)
         is_expected.to forbid_action(:destroy)
         is_expected.to permit_action(:redact_from_iati)
         is_expected.to forbid_action(:create_refund)
@@ -92,10 +92,10 @@ RSpec.describe ActivityPolicy do
     context "when the activity is a project" do
       let(:activity) { create(:project_activity) }
 
-      it "only permits show, redact_from_iati, download and set_commitment" do
+      it "only permits show, redact_from_iati, show_xml and set_commitment" do
         is_expected.to permit_action(:show)
         is_expected.to permit_action(:redact_from_iati)
-        is_expected.to permit_action(:download)
+        is_expected.to permit_action(:show_xml)
 
         is_expected.to forbid_action(:create_refund)
         is_expected.to forbid_action(:create_adjustment)
@@ -149,10 +149,10 @@ RSpec.describe ActivityPolicy do
     context "when the activity is a third-party project" do
       let(:activity) { create(:third_party_project_activity) }
 
-      it "only permits show, redact_from_iati, download and set_commitment" do
+      it "only permits show, redact_from_iati, show_xml and set_commitment" do
         is_expected.to permit_action(:show)
         is_expected.to permit_action(:redact_from_iati)
-        is_expected.to permit_action(:download)
+        is_expected.to permit_action(:show_xml)
 
         is_expected.to forbid_action(:create)
         is_expected.to forbid_action(:edit)
@@ -207,7 +207,7 @@ RSpec.describe ActivityPolicy do
         is_expected.to forbid_action(:edit)
         is_expected.to forbid_action(:update)
         is_expected.to forbid_action(:destroy)
-        is_expected.to forbid_action(:download)
+        is_expected.to forbid_action(:show_xml)
         is_expected.to forbid_action(:redact_from_iati)
 
         is_expected.to forbid_action(:create_child)
@@ -229,7 +229,7 @@ RSpec.describe ActivityPolicy do
           is_expected.to forbid_action(:edit)
           is_expected.to forbid_action(:update)
           is_expected.to forbid_action(:destroy)
-          is_expected.to forbid_action(:download)
+          is_expected.to forbid_action(:show_xml)
           is_expected.to forbid_action(:redact_from_iati)
 
           is_expected.to forbid_action(:create_child)
@@ -253,7 +253,7 @@ RSpec.describe ActivityPolicy do
           is_expected.to forbid_action(:edit)
           is_expected.to forbid_action(:update)
           is_expected.to forbid_action(:destroy)
-          is_expected.to forbid_action(:download)
+          is_expected.to forbid_action(:show_xml)
           is_expected.to forbid_action(:redact_from_iati)
 
           is_expected.to forbid_action(:create_child)
@@ -291,7 +291,7 @@ RSpec.describe ActivityPolicy do
           is_expected.to forbid_action(:edit)
           is_expected.to forbid_action(:update)
           is_expected.to forbid_action(:destroy)
-          is_expected.to forbid_action(:download)
+          is_expected.to forbid_action(:show_xml)
           is_expected.to forbid_action(:redact_from_iati)
 
           is_expected.to forbid_action(:create_child)
@@ -320,7 +320,7 @@ RSpec.describe ActivityPolicy do
             is_expected.to forbid_action(:edit)
             is_expected.to forbid_action(:update)
             is_expected.to forbid_action(:destroy)
-            is_expected.to forbid_action(:download)
+            is_expected.to forbid_action(:show_xml)
             is_expected.to forbid_action(:redact_from_iati)
 
             is_expected.to forbid_action(:create_child)
@@ -337,13 +337,13 @@ RSpec.describe ActivityPolicy do
             report.update(state: :active)
           end
 
-          it "only forbids download, destroy, redact_from_iati, and update_linked_activity" do
+          it "only forbids show_xml, destroy, redact_from_iati, and update_linked_activity" do
             is_expected.to permit_action(:show)
             is_expected.to permit_action(:create)
             is_expected.to permit_action(:edit)
             is_expected.to permit_action(:update)
 
-            is_expected.to forbid_action(:download)
+            is_expected.to forbid_action(:show_xml)
             is_expected.to forbid_action(:destroy)
             is_expected.to forbid_action(:redact_from_iati)
 
@@ -391,7 +391,7 @@ RSpec.describe ActivityPolicy do
           is_expected.to forbid_action(:create)
           is_expected.to forbid_action(:edit)
           is_expected.to forbid_action(:update)
-          is_expected.to forbid_action(:download)
+          is_expected.to forbid_action(:show_xml)
           is_expected.to forbid_action(:destroy)
           is_expected.to forbid_action(:redact_from_iati)
 
@@ -416,7 +416,7 @@ RSpec.describe ActivityPolicy do
             is_expected.to forbid_action(:create)
             is_expected.to forbid_action(:edit)
             is_expected.to forbid_action(:update)
-            is_expected.to forbid_action(:download)
+            is_expected.to forbid_action(:show_xml)
             is_expected.to forbid_action(:destroy)
             is_expected.to forbid_action(:redact_from_iati)
 
@@ -433,13 +433,13 @@ RSpec.describe ActivityPolicy do
             report.update(state: :active)
           end
 
-          it "only forbids download, destroy, redact_from_iati, create_child, and update_linked_activity" do
+          it "only forbids show_xml, destroy, redact_from_iati, create_child, and update_linked_activity" do
             is_expected.to permit_action(:show)
             is_expected.to permit_action(:create)
             is_expected.to permit_action(:edit)
             is_expected.to permit_action(:update)
 
-            is_expected.to forbid_action(:download)
+            is_expected.to forbid_action(:show_xml)
             is_expected.to forbid_action(:destroy)
             is_expected.to forbid_action(:redact_from_iati)